recently blank でファイルがない場合画面が表示されない
