+++ src/store/index.ts
@@ -5,8 +5,8 @@ import { INITIAL_NOTE, VIEW_MODE } from '@/constants'
 import { updateBrowsingHistory } from '@/utils/local-storage'
 import Editor from '@/assets/scripts/editor/markdown-editor'
 import { Preference } from '@/config/setting'
-import { getPreference, updatePreference } from '@/utils/local-storage'
-import { getBrowsingHistories } from '@/utils/local-storage'
+import { getPreference, updatePreference, getBrowsingHistories } from '@/utils/local-storage'
+import { isExistPath } from '@/utils/note'
 
 Vue.use(Vuex)
 
@@ -27,7 +27,10 @@ interface State {
 const preference: Preference = getPreference()
 let path
 if (preference.initialNote === INITIAL_NOTE.RECENTRY_OPENED) {
-  path = getBrowsingHistories()[0]
+  const latestPath = getBrowsingHistories()[0]
+  if (isExistPath(latestPath)) {
+    path = latestPath
+  }
 }
 const note: Note = new Note(path)
